Version 1.0.6.0
use of self.FastvsFastTag and self.FastvsFullTag.
Usually, the Tag for FastvsFull was extracted from FastvsFast (we only
removed « _FastSim » from Tag). With time, another way was used,
introducing « realistic » for FullSim and « asymptotic » for FastSim.
This imply to modify « by hand » the genberated html files.
A modification was introduiced : a FastSim line (red) appear only for
FastSim and permit the Tag modification.by default the presented one is
from FastvsFast. These modifications and the fact that that keeping the
actual release (instead of the reference) for the FastvsFull are
integrated into OvalFile.